+++ dialog_alter/main.cpp
@@ -1,0 +1,15 @@
+#include "main_window.h"
+#include <QApplication>
+#include "my_dialog.h"
+int main(int argc, char *argv[])
+{
+    QApplication a(argc, argv);
+    MainWindow w;
+    MyDialog dialog;
+    if(dialog.exec() == QDialog::Accepted){
+         w.show();
+         return a.exec();
+    }else{
+        return 0;
+    }
+}