+++ dialog_alter/main_window.cpp
@@ -1,0 +1,26 @@
+#include "main_window.h"
+#include "ui_main_window.h"
+#include "my_dialog.h"
+MainWindow::MainWindow(QWidget *parent) :
+    QDialog(parent),
+    ui(new Ui::MainWindow)
+{
+    ui->setupUi(this);
+}
+
+MainWindow::~MainWindow()
+{
+    delete ui;
+}
+
+void MainWindow::on_pushButton_clicked()
+{
+    close();
+    MyDialog mydialog;
+    if(mydialog.exec() == QDialog::Accepted ){
+        this->show();
+    }else{
+        return;
+    }
+}
+