+++ dialog_alter/main_window.h
@@ -1,0 +1,25 @@
+#ifndef MAIN_WINDOW_H
+#define MAIN_WINDOW_H
+
+#include <QDialog>
+
+namespace Ui {
+class MainWindow;
+}
+
+class MainWindow : public QDialog
+{
+    Q_OBJECT
+
+public:
+    explicit MainWindow(QWidget *parent = nullptr);
+    ~MainWindow();
+
+private slots:
+    void on_pushButton_clicked();
+
+private:
+    Ui::MainWindow *ui;
+};
+
+#endif // MAIN_WINDOW_H