+++ rich_texter/main_window.h
@@ -1,0 +1,23 @@
+#ifndef MAIN_WINDOW_H
+#define MAIN_WINDOW_H
+
+#include <QMainWindow>
+class MySyntaxHighlighter;
+namespace Ui {
+class MainWindow;
+}
+
+class MainWindow : public QMainWindow
+{
+    Q_OBJECT
+
+public:
+    explicit MainWindow(QWidget *parent = nullptr);
+    ~MainWindow();
+
+private:
+    Ui::MainWindow *ui;
+    MySyntaxHighlighter * m_ligher;
+};
+
+#endif // MAIN_WINDOW_H