+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # 记录
-* QT 主窗口
+* 按键事件和定时器
+
 # 2026-04-27
 * QT 布局
+* QT 主窗口
+* 自定义菜单项
+* QTextEdit
+* 文本高亮
+* QDragEvent
+* Event Filter
+* 鼠标事件
 # 2026-04-26
 * 学完QT对话框
 * QT LineEdit